Hash debtor/creditorAccount's IBAN number before lookup
