+++ packages/loot-core/src/server/encryption-internals.js
@@ -2,10 +2,6 @@ import crypto from 'crypto';
 
 let ENCRYPTION_ALGORITHM = 'aes-256-gcm';
 
-export async function sha256String(str) {
-  return crypto.createHash('sha256').update(str).digest('base64');
-}
-
 export function randomBytes(n) {
   return crypto.randomBytes(n);
 }
