+++ packages/loot-core/src/server/util/hash.js
@@ -1,0 +1,5 @@
+let crypto = require('crypto');
+
+export async function sha256String(str) {
+  return crypto.createHash('sha256').update(str).digest('base64');
+}
